requirements.txt
django==1.10
django-allauth==0.31.0
Pillow==4.3.0